pyproject.toml
[tool.poetry]
name = "discordunmute"
version = "0.1.0"
description = "指定VCから抜けたら自動的にミュートを外します"
authors = ["Kamesuta <kamesuta@gmail.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11.1"
"discord.py" = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
